requirements.txt
torch>=2.3
transformers>=4.44
trl>=0.11
peft>=0.12
accelerate>=0.33
bitsandbytes>=0.43 ; platform_system != "Darwin"
datasets>=2.20
pytest>=8
coverage>=7.6
openai>=1.40
pyyaml
python-dotenv
tqdm
matplotlib
wandb
